refactor[spring-jdbc]: refactor author, genre, book functionality

+++ bookstore-spring-jdbc/src/main/java/com/github/ducknowledges/bookstore/dao/AuthorDaoJdbc.java
@@ -21,7 +21,7 @@ public class AuthorDaoJdbc implements AuthorDao {
     }
 
     @Override
-    public Optional<Author> readById(int id) {
+    public Optional<Author> readById(long id) {
         Author author;
         try {
             author = jdbc.queryForObject("select id, name from author where id = :id",
