+++ bookstore-spring-jdbc/src/main/java/com/github/ducknowledges/bookstore/dao/mapper/BookRowMapper.java
@@ -12,10 +12,10 @@ import org.springframework.stereotype.Component;
 public class BookRowMapper implements RowMapper<Book> {
     @Override
     public Book mapRow(ResultSet rs, int rowNumber) throws SQLException {
-        int id = rs.getInt("id");
+        long id = rs.getLong("id");
         String name = rs.getString("name");
-        Author author = new Author(rs.getInt("author_id"), rs.getString("author_name"));
-        Genre genre = new Genre(rs.getInt("author_id"), rs.getString("genre_name"));
+        Author author = new Author(rs.getLong("author_id"), rs.getString("author_name"));
+        Genre genre = new Genre(rs.getLong("author_id"), rs.getString("genre_name"));
         return new Book(id, name, author, genre);
     }
 }
