+++ bookstore-spring-jdbc/src/main/resources/db/migration/schema/1.0/V1_1__2023_01_19_Add_author_table.sql
@@ -3,6 +3,6 @@
 
 create table author
 (
-    id   int auto_increment primary key,
+    id   bigint auto_increment primary key,
     name varchar(255) not null unique
 );
