+++ bookstore-spring-jdbc/src/main/resources/db/migration/schema/2.0/V2_1__2023_01_21_Add_book_table.sql
@@ -1,6 +1,6 @@
 create table book
 (
-    id        int auto_increment primary key,
+    id        bigint auto_increment primary key,
     name      varchar(255) not null,
     author_id int          not null,
     genre_id  int          not null
